Consolidated test logic
